finish following video

+++ pantry/app/page.js
@@ -2,7 +2,7 @@
 
 import {Box, Button, Modal, Stack, TextField, Typography} from '@mui/material'
 import { firestore } from '@/firebase';
-import { collection, getDocs, query, setDoc, doc, deleteDoc } from 'firebase/firestore';
+import { collection, getDocs, query, setDoc, doc, deleteDoc, getDoc} from 'firebase/firestore';
 import { useEffect, useState } from 'react';
 // 33:49
 const style = {
@@ -34,7 +34,7 @@ export default function Home() {
     const docs = await getDocs(snapshot)
     const pantryList = []
     docs.forEach((doc) => {
-      pantryList.push(doc.id)
+      pantryList.push({name: doc.id, ...doc.data()})
     })
     console.log(pantryList)
     setPantry(pantryList)
@@ -46,13 +46,29 @@ export default function Home() {
   
   const addItem = async (item) => {
     const docRef = doc(collection(firestore, 'pantry'), item)
-    await setDoc(docRef, {})
+    const docSnap = await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      await setDoc(docRef, {count: count + 1})
+    }
+    else {
+      await setDoc(docRef, {count: 1})
+    }
     await updatePantry()
   }
 
   const removeItem = async (item) => {
     const docRef = doc(collection(firestore, 'pantry'), item)
-    await deleteDoc(docRef)
+    const docSnap = await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      if (count === 1) {
+        await deleteDoc(docRef)
+      }
+      else {
+        await setDoc(docRef, {count: count - 1})
+      }
+    }
     await updatePantry()
   }
 
@@ -99,9 +115,9 @@ export default function Home() {
           height="300px"
           spacing={2}
           overflow={'auto'}>
-            {pantry.map((i) => (
+            {pantry.map(({name, count}) => (
                 <Box
-                  key={i}
+                  key={name}
                   width="100%"
                   minHeight="150px"
                   display={'flex'}
@@ -111,10 +127,15 @@ export default function Home() {
                   paddingX={5}>
                   <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
                       {
-                        i.charAt(0).toUpperCase() + i.slice(1)
+                        name.charAt(0).toUpperCase() + name.slice(1)
                       }
                   </Typography>
-                  <Button variant="contained" onClick={() => removeItem(i)}>Remove</Button>
+
+                  <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
+                      Quantity: {count}
+                  </Typography>
+
+                  <Button variant="contained" onClick={() => removeItem(name)}>Remove</Button>
                 </Box>
             ))}
         </Stack>
